[REFACTOR-017] extract guide-types, BlockEditor, CoverImageUploader; fix import paths
