+++ app/admin/content/components/BlockEditor.tsx
@@ -1,0 +1,114 @@
+'use client'
+
+import { useLanguage } from '@/lib/hooks/useLanguage'
+import type { Block } from './guide-types'
+
+export function BlockEditor({
+  blocks,
+  onChange,
+}: {
+  blocks: Block[]
+  onChange: (b: Block[]) => void
+}): React.JSX.Element {
+  const { t } = useLanguage()
+  const safeBlocks = Array.isArray(blocks) ? blocks : []
+
+  function addBlock(type: Block['type']) {
+    onChange([...safeBlocks, { type, content: { en: '', bg: '' }, emoji: type === 'callout' ? '\uD83D\uDCA1' : undefined }])
+  }
+  function updateBlock(i: number, partial: Partial<Block>) {
+    onChange(safeBlocks.map((b, idx) => idx === i ? { ...b, ...partial } : b))
+  }
+  function updateContent(i: number, lang: 'en' | 'bg', value: string) {
+    updateBlock(i, { content: { ...safeBlocks[i].content, [lang]: value } })
+  }
+  function moveBlock(i: number, dir: -1 | 1) {
+    const next = [...safeBlocks]
+    const j = i + dir
+    if (j < 0 || j >= next.length) return
+    ;[next[i], next[j]] = [next[j], next[i]]
+    onChange(next)
+  }
+  function removeBlock(i: number) {
+    onChange(safeBlocks.filter((_, idx) => idx !== i))
+  }
+
+  return (
+    <div className="space-y-3">
+      {safeBlocks.map((block, i) => (
+        <div key={i} className="rounded-xl border p-4 space-y-3"
+          style={{ backgroundColor: 'var(--bg-card)', borderColor: 'var(--border-default)' }}>
+          <div className="flex items-center justify-between gap-2">
+            <div className="flex items-center gap-2">
+              <span className="text-xs font-semibold px-2 py-0.5 rounded-full uppercase tracking-widest"
+                style={{
+                  backgroundColor: block.type === 'heading' ? 'var(--brand-forest)' : block.type === 'callout' ? 'var(--brand-teal)' : 'rgba(0,0,0,0.06)',
+                  color: block.type === 'paragraph' ? 'var(--text-secondary)' : 'var(--brand-parchment)',
+                }}>
+                {block.type}
+              </span>
+              {block.type === 'callout' && (
+                <input
+                  value={block.emoji ?? ''}
+                  onChange={e => updateBlock(i, { emoji: e.target.value })}
+                  placeholder="emoji"
+                  className="w-14 border rounded-lg px-2 py-1 text-sm text-center"
+                  style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)' }}
+                />
+              )}
+            </div>
+            <div className="flex items-center gap-1">
+              <button onClick={() => moveBlock(i, -1)} disabled={i === 0}
+                className="w-6 h-6 flex items-center justify-center rounded hover:bg-black/5 disabled:opacity-30 text-xs"
+                style={{ color: 'var(--text-secondary)' }}>↑</button>
+              <button onClick={() => moveBlock(i, 1)} disabled={i === safeBlocks.length - 1}
+                className="w-6 h-6 flex items-center justify-center rounded hover:bg-black/5 disabled:opacity-30 text-xs"
+                style={{ color: 'var(--text-secondary)' }}>↓</button>
+              <button onClick={() => removeBlock(i)}
+                className="w-6 h-6 flex items-center justify-center rounded hover:bg-black/5 text-xs"
+                style={{ color: 'var(--brand-crimson)' }}>✕</button>
+            </div>
+          </div>
+          <div className="grid grid-cols-2 gap-3">
+            {(['en', 'bg'] as const).map(lang => (
+              <div key={lang}>
+                <label className="text-[10px] font-semibold uppercase tracking-widest mb-1 block"
+                  style={{ color: 'var(--text-secondary)' }}>{lang.toUpperCase()}</label>
+                {block.type === 'paragraph' || block.type === 'callout' ? (
+                  <>
+                    <textarea
+                      value={block.content[lang]}
+                      onChange={e => updateContent(i, lang, e.target.value)}
+                      rows={3}
+                      className="w-full border rounded-xl px-3 py-2 text-sm resize-none"
+                      style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }}
+                    />
+                    <p className="text-[10px] mt-1" style={{ color: 'var(--text-tertiary)' }}>
+                      {t('admin.content.guides.markdownHint')}
+                    </p>
+                  </>
+                ) : (
+                  <input
+                    value={block.content[lang]}
+                    onChange={e => updateContent(i, lang, e.target.value)}
+                    className="w-full border rounded-xl px-3 py-2 text-sm"
+                    style={{ borderColor: 'var(--border-default)', color: 'var(--text-primary)', backgroundColor: 'var(--bg-card)' }}
+                  />
+                )}
+              </div>
+            ))}
+          </div>
+        </div>
+      ))}
+      <div className="flex gap-2 pt-1">
+        {(['heading', 'paragraph', 'callout'] as const).map(type => (
+          <button key={type} onClick={() => addBlock(type)}
+            className="px-3 py-1.5 rounded-lg text-xs font-semibold border transition-colors hover:bg-black/5"
+            style={{ borderColor: 'var(--border-default)', color: 'var(--text-secondary)' }}>
+            + {type}
+          </button>
+        ))}
+      </div>
+    </div>
+  )
+}
