+++ app/admin/content/components/guide-types.ts
@@ -1,0 +1,41 @@
+// Shared domain types and utilities for guide/news content.
+// No 'use client' — plain TypeScript module.
+
+export type Block = {
+  type: 'heading' | 'paragraph' | 'callout'
+  content: { en: string; bg: string }
+  emoji?: string
+}
+
+export type Guide = {
+  id: string
+  slug: string
+  title: { en: string; bg: string }
+  cover_image_url: string | null
+  emoji: string | null
+  body: Block[]
+  access_roles: string[]
+  is_published: boolean
+  created_at: string
+  updated_at: string
+  sort_order: number
+}
+
+export const ALL_ROLES = ['guest', 'member', 'core', 'admin']
+
+export function slugify(str: string): string {
+  return str.toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/(^-|-$)/g, '')
+}
+
+export function emptyGuide(): Omit<Guide, 'id' | 'created_at' | 'updated_at'> {
+  return {
+    slug: '',
+    title: { en: '', bg: '' },
+    cover_image_url: null,
+    emoji: null,
+    body: [],
+    access_roles: [...ALL_ROLES],
+    is_published: false,
+    sort_order: 0,
+  }
+}
